src/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0 
python-multipart==0.0.20 
python-dotenv==1.0.1
pydantic-settings==2.7.0
aiofiles==24.1.0 
langchain==0.3.13
langchain-community==0.3.13
PyMuPDF==1.25.1
motor==3.6.0


